+++ CHANGELOG.md
@@ -87,7 +87,16 @@
   `RTTI` ([#604](https://github.com/odygrd/quill/issues/604))
 - Fixed an incorrectly triggered assertion in debug builds when `BackendOptions::log_timestamp_ordering_grace_period` is
   set to 0 ([#605](https://github.com/odygrd/quill/issues/605))
-- Fixed a compile-time error in `CsvWriter` that occurred when passing a custom `FrontendOptions` type as a template parameter. ([#609](https://github.com/odygrd/quill/issues/609))
+- Fixed a compile-time error in `CsvWriter` that occurred when passing a custom `FrontendOptions` type as a template
+  parameter. ([#609](https://github.com/odygrd/quill/issues/609))
+- Added `ConsoleColours::ColourMode` to `ConsoleSink`, allowing colors to be explicitly forced or conditionally enabled
+  based on the environment. Previously, colors were only conditionally
+  enabled. ([#611](https://github.com/odygrd/quill/issues/611)).
+
+  For example:
+  ```cpp
+    quill::Frontend::create_or_get_sink<quill::ConsoleSink>(
+      "sink_id_1", quill::ConsoleColours::ColourMode::Automatic);
 
 ## v7.3.0